Split qml.go into lots of files
